refactor: consolidate date utils and enforce kebab-case lib naming
Merge three overlapping date utility files (date-utils.ts, dateUtils.ts,
dates/streakDateUtils.ts) into a single src/lib/date-utils.ts. Rename
camelCase type file repoAnalytics.ts → repo-analytics-types.ts and remove
the thin repoAnalyticsUtils.ts re-export wrapper. Update all 21 affected
import sites.
