+++ src/app/api/metrics/repo-analytics/route.ts
@@ -3,7 +3,7 @@ import { NextRequest } from "next/server";
 import { authOptions } from "@/lib/auth";
 import { isMetricsCacheBypassed, metricsCacheKey, withMetricsCache } from "@/lib/metrics-cache";
 import { computeHealthScore } from "@/lib/repo-health";
-import { RepoAnalyticsResponse } from "@/lib/repoAnalytics";
+import { RepoAnalyticsResponse } from "@/lib/repo-analytics-types";
 import { isSafeUrl } from "@/lib/ssrf-protection";
 import { parseRepoParam } from "@/lib/repo-analytics-utils";
