+++ src/app/api/metrics/repo-explorer/route.ts
@@ -1,7 +1,7 @@
 import { getSessionWithToken } from "@/lib/get-session-token";
 import { NextRequest } from "next/server";
 import { isMetricsCacheBypassed, metricsCacheKey, withMetricsCache } from "@/lib/metrics-cache";
-import { ExplorerRepoCardData } from "@/lib/repoAnalytics";
+import { ExplorerRepoCardData } from "@/lib/repo-analytics-types";
 
 export const dynamic = "force-dynamic";
 const GITHUB_API = "https://api.github.com";
